feat: scrollbar util 함수로 분리
